PIM-4980: move format providers in a specific folder
