+++ src/Pim/Bundle/LocalizationBundle/Resources/config/providers.yml
@@ -1,6 +1,7 @@
 parameters:
     pim_localization.provider.ui_locale.class:          Pim\Bundle\LocalizationBundle\Provider\UiLocaleProvider
-    pim_localization.provider.date_format.class:        Pim\Component\Localization\Provider\DateFormatProvider
+    pim_localization.provider.date_format.class:        Pim\Component\Localization\Provider\Format\DateFormatProvider
+    pim_localization.provider.number_format.class:      Pim\Component\Localization\Provider\Format\NumberFormatProvider
     pim_localization.provider.ui_locale.min_percentage: 0.8
     pim_localization.provider.date_format.formats:
           en_US: 'n/j/y'
@@ -18,3 +19,6 @@ services:
         arguments:
             - '@pim_localization.date_format_converter'
             - %pim_localization.provider.date_format.formats%
+
+    pim_localization.provider.number_format:
+        class: %pim_localization.provider.number_format.class%
